Add trigger to keep maximum of 10 posts
New Postgres trigger function aims to keep a maximum of 10 rows in the posts table at any one time. Before a new row is inserted, the oldest rows will be deleted until 9 rows remain; the new row will bring the total count back up to 10.

I could increase the number of allowed posts in the future, but I want some sort of limit, so that the database won't grow in size infinitely.