Add `getPosts` route, alter `ListPosts` query
The route returns posts from the database. The query for this has been altered so that a maximum of 10 posts are returned. The trigger function from the previous commit should ensure that there are only ever at most 10 posts in the table, but I just want to be more sure that the `getPosts` route won't end up returning tons of data unexpectedly.