Fix bug when emoji submission fails
React seems to be resetting the form even if submission fails. This is causing some odd behaviour because I have a controlled component in the form - the <select> element. Resetting the state value used by the <select> element seems to improve things: the form is now properly reset even when form submission fails.