Remove Parcel, add Vite instead
React's docs for the React Compiler mention Vite but not Parcel. It might be easier to enable the React Compiler with Vite. Also Parcel doesn't seem very fast on my machine.

Vite required me to move a couple of files (index.html and index.tsx), and I had to change how env variables are handled.