Add possibly unnecessary optimisation
In the last commit, PostsList was re-rendering every second. I have refactored the code so that now it seems only the Countdown component is re-rendering every second. I tested the performance with the React Profiler (part of the React Developer Tools browser extension). Before, each render seemed to take at least 1.4 ms. Now, each render seems to take about 0.4 - 0.6 ms. Perhaps this optimisation is small and unnecessary, but hopefully it makes this website a little more efficient.

I have replaced `setInterval` with `setTimeout`. I don't think there is much point to `setInterval` in this new code, because the effect has to run every second anyway, because it is dependent upon the value of `secondsElapsed`. I tried getting the `secondsElapsed` value within the setter, `setSecondsElapsed`, but then you need to set `dateUpdatedEveryMinute` within the `setSecondsElapsed` function, and React doesn't seem to like that - it will show this error:

"Cannot update a component (`PostsList`) while rendering a different component (`Countdown`). To locate the bad setState() call inside `Countdown`, follow the stack trace as described in https://react.dev/link/setstate-in-render"