Add migration to fix race condition
When sending lots of POST requests to /api/v1/posts in quick succession, the database produced unwanted behaviour. It could sometimes end up with 13 rows in the posts table, which is unwanted. I want there to always be a maximum of 10 posts in the posts table.

The migration in this commit should fix this issue. I'm using a lock on the posts table which seems to ensure that after an insert (but before the transaction completes), any additional posts beyond the newest 10 posts are deleted.

There is an issue with this locking. It means that when lots of inserts are beyond performed, you can't load posts from the database. So if one user on the internet is spamming the back-end with POST requests to /api/v1/posts, other users won't be able to load posts at all - loading will just hang. Perhaps I should add rate limiting so that the back-end can't be spammed with POST requests in this way.